+++ nebula_decoders/src/nebula_decoders_velodyne/decoders/vlp32_decoder.cpp
@@ -133,7 +133,7 @@ void Vlp32Decoder::reset_overflow(double time_stamp)
   overflow_pc_->points.reserve(max_pts_);
 }
 
-void Vlp32Decoder::unpack(const std::vector<uint8_t> & packet, int32_t packet_seconds)
+void Vlp32Decoder::unpack(const std::vector<uint8_t> & packet, double packet_seconds)
 {
   checkAndHandleScanComplete(packet, packet_seconds, phase_);
