+++ nebula_decoders/src/nebula_decoders_velodyne/decoders/vls128_decoder.cpp
@@ -135,7 +135,7 @@ void Vls128Decoder::reset_overflow(double time_stamp)
   overflow_pc_->points.reserve(max_pts_);
 }
 
-void Vls128Decoder::unpack(const std::vector<uint8_t> & packet, int32_t packet_seconds)
+void Vls128Decoder::unpack(const std::vector<uint8_t> & packet, double packet_seconds)
 {
   checkAndHandleScanComplete(packet, packet_seconds, phase_);
